backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.37
asyncpg==0.29.0
redis==5.2.1
rq==1.16.2
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.27.2
ccxt==4.3.87
